Make these account tests not slow (#5612)
