+++ ironfish/src/rpc/routes/wallet/create.test.ts
@@ -9,9 +9,9 @@ import { v4 as uuid } from 'uuid'
 import { createRouteTest } from '../../../testUtilities/routeTest'
 import { RPC_ERROR_CODES } from '../../adapters'
 import { RpcRequestError } from '../../clients/errors'
+import { CreateAccountResponse } from './createAccount'
 
-describe('Route wallet/createAccount', () => {
-  jest.setTimeout(15000)
+describe('Route wallet/create', () => {
   const routeTest = createRouteTest()
 
   beforeEach(() => {
@@ -27,7 +27,8 @@ describe('Route wallet/createAccount', () => {
 
     const name = uuid()
 
-    const response = await routeTest.client.wallet.createAccount({ name })
+    const response = await routeTest.client.request(`wallet/create`, { name }).waitForEnd()
+
     expect(response.status).toBe(200)
     expect(response.content).toMatchObject({
       name: name,
@@ -38,7 +39,7 @@ describe('Route wallet/createAccount', () => {
     const account = routeTest.node.wallet.getAccountByName(name)
     expect(account).toMatchObject({
       name: name,
-      publicAddress: response.content.publicAddress,
+      publicAddress: (response.content as CreateAccountResponse).publicAddress,
       createdAt: createdAtHead,
     })
   })
@@ -48,7 +49,8 @@ describe('Route wallet/createAccount', () => {
 
     const name = uuid()
 
-    const response = await routeTest.client.wallet.createAccount({ name })
+    const response = await routeTest.client.request(`wallet/create`, { name }).waitForEnd()
+
     expect(response.content).toMatchObject({
       name: name,
       publicAddress: expect.any(String),
@@ -64,13 +66,13 @@ describe('Route wallet/createAccount', () => {
 
     try {
       expect.assertions(2)
-      await routeTest.client.wallet.createAccount({ name: name })
+      await routeTest.client.request(`wallet/create`, { name }).waitForEnd()
     } catch (e: unknown) {
       if (!(e instanceof RpcRequestError)) {
         throw e
       }
       expect(e.status).toBe(400)
-      expect(e.code).toBe(RPC_ERROR_CODES.DUPLICATE_ACCOUNT_NAME)
+      expect(e.code).toBe(RPC_ERROR_CODES.ACCOUNT_EXISTS)
     }
   })
 
@@ -83,7 +85,8 @@ describe('Route wallet/createAccount', () => {
 
     const name = uuid()
 
-    const response = await routeTest.client.wallet.createAccount({ name })
+    const response = await routeTest.client.request(`wallet/create`, { name }).waitForEnd()
+
     expect(response.status).toBe(200)
     expect(response.content).toMatchObject({
       name: name,
@@ -92,55 +95,5 @@ describe('Route wallet/createAccount', () => {
     })
 
     expect(scanSpy).toHaveBeenCalled()
-  })
-
-  it('should set account createdAt if passed', async () => {
-    const name = uuid()
-    const createdAt = 10
-
-    const response = await routeTest.client.wallet.createAccount({
-      name,
-      createdAt,
-    })
-
-    expect(response.status).toBe(200)
-    expect(response.content).toMatchObject({
-      name: name,
-      publicAddress: expect.any(String),
-      isDefaultAccount: true,
-    })
-
-    const account = routeTest.node.wallet.getAccountByName(name)
-    expect(account).toMatchObject({
-      name: name,
-      publicAddress: response.content.publicAddress,
-      createdAt: {
-        hash: Buffer.alloc(32, 0),
-        sequence: 10,
-      },
-    })
-  })
-
-  it('should set account createdAt to null', async () => {
-    const name = uuid()
-
-    const response = await routeTest.client.wallet.createAccount({
-      name,
-      createdAt: null,
-    })
-
-    expect(response.status).toBe(200)
-    expect(response.content).toMatchObject({
-      name: name,
-      publicAddress: expect.any(String),
-      isDefaultAccount: true,
-    })
-
-    const account = routeTest.node.wallet.getAccountByName(name)
-    expect(account).toMatchObject({
-      name: name,
-      publicAddress: response.content.publicAddress,
-      createdAt: null,
-    })
   })
 })
