+++ ironfish/src/rpc/routes/wallet/createAccount.test.ts
@@ -9,10 +9,8 @@ import { v4 as uuid } from 'uuid'
 import { createRouteTest } from '../../../testUtilities/routeTest'
 import { RPC_ERROR_CODES } from '../../adapters'
 import { RpcRequestError } from '../../clients/errors'
-import { CreateAccountResponse } from './createAccount'
 
-describe('Route wallet/create', () => {
-  jest.setTimeout(15000)
+describe('Route wallet/createAccount', () => {
   const routeTest = createRouteTest()
 
   beforeEach(() => {
@@ -28,8 +26,7 @@ describe('Route wallet/create', () => {
 
     const name = uuid()
 
-    const response = await routeTest.client.request(`wallet/create`, { name }).waitForEnd()
-
+    const response = await routeTest.client.wallet.createAccount({ name })
     expect(response.status).toBe(200)
     expect(response.content).toMatchObject({
       name: name,
@@ -40,7 +37,7 @@ describe('Route wallet/create', () => {
     const account = routeTest.node.wallet.getAccountByName(name)
     expect(account).toMatchObject({
       name: name,
-      publicAddress: (response.content as CreateAccountResponse).publicAddress,
+      publicAddress: response.content.publicAddress,
       createdAt: createdAtHead,
     })
   })
@@ -50,8 +47,7 @@ describe('Route wallet/create', () => {
 
     const name = uuid()
 
-    const response = await routeTest.client.request(`wallet/create`, { name }).waitForEnd()
-
+    const response = await routeTest.client.wallet.createAccount({ name })
     expect(response.content).toMatchObject({
       name: name,
       publicAddress: expect.any(String),
@@ -67,13 +63,13 @@ describe('Route wallet/create', () => {
 
     try {
       expect.assertions(2)
-      await routeTest.client.request(`wallet/create`, { name }).waitForEnd()
+      await routeTest.client.wallet.createAccount({ name: name })
     } catch (e: unknown) {
       if (!(e instanceof RpcRequestError)) {
         throw e
       }
       expect(e.status).toBe(400)
-      expect(e.code).toBe(RPC_ERROR_CODES.ACCOUNT_EXISTS)
+      expect(e.code).toBe(RPC_ERROR_CODES.DUPLICATE_ACCOUNT_NAME)
     }
   })
 
@@ -86,8 +82,7 @@ describe('Route wallet/create', () => {
 
     const name = uuid()
 
-    const response = await routeTest.client.request(`wallet/create`, { name }).waitForEnd()
-
+    const response = await routeTest.client.wallet.createAccount({ name })
     expect(response.status).toBe(200)
     expect(response.content).toMatchObject({
       name: name,
@@ -96,5 +91,55 @@ describe('Route wallet/create', () => {
     })
 
     expect(scanSpy).toHaveBeenCalled()
+  })
+
+  it('should set account createdAt if passed', async () => {
+    const name = uuid()
+    const createdAt = 10
+
+    const response = await routeTest.client.wallet.createAccount({
+      name,
+      createdAt,
+    })
+
+    expect(response.status).toBe(200)
+    expect(response.content).toMatchObject({
+      name: name,
+      publicAddress: expect.any(String),
+      isDefaultAccount: true,
+    })
+
+    const account = routeTest.node.wallet.getAccountByName(name)
+    expect(account).toMatchObject({
+      name: name,
+      publicAddress: response.content.publicAddress,
+      createdAt: {
+        hash: Buffer.alloc(32, 0),
+        sequence: 10,
+      },
+    })
+  })
+
+  it('should set account createdAt to null', async () => {
+    const name = uuid()
+
+    const response = await routeTest.client.wallet.createAccount({
+      name,
+      createdAt: null,
+    })
+
+    expect(response.status).toBe(200)
+    expect(response.content).toMatchObject({
+      name: name,
+      publicAddress: expect.any(String),
+      isDefaultAccount: true,
+    })
+
+    const account = routeTest.node.wallet.getAccountByName(name)
+    expect(account).toMatchObject({
+      name: name,
+      publicAddress: response.content.publicAddress,
+      createdAt: null,
+    })
   })
 })
